Open cgroup only when necessary
This prevents a cgroup from being created when it is not necessary. For
example, a regular box without CPU or memory restriction should not have
to create a cgroup. Likewise, when the `CI_NO_CGROUP` environment
variable is set, we can skip opening a cgroup.

This will allow running foxbox in situations where we can't use cgroups
or they aren't available yet (like #7).